requirements.txt
alembic==1.5.7
apispec==4.3.0
attrs==20.3.0
click==7.1.2
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
greenlet==1.0.0
hashids==1.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
mistune==0.8.4
psycopg2==2.8.4
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.4.0
validators==0.18.2
Werkzeug==1.0.1
